Move operations for fixing into an object
